+++ scripts/node/icons/icons.config.js
@@ -13,23 +13,28 @@ const PATH_ICONS_LIB = `${LIB_ROOT}/icons/`;
 module.exports = {
   // Options for exporting icons from sketch
   sketchConfig: {
-    output: `${ICONS_ROOT}/svg`,
     pageName: 'Icons',
     slicePrefix: 'icon/',
   },
 
+  // paths used by build scripts
   buildConfig: {
-    raw: {
-      input: `${ICONS_ROOT}/svg`,
-      output: `${PATH_ICONS_LIB}/raw`,
+    svg: {
+      src: `${ICONS_ROOT}/svg`,
+      lib: `${PATH_ICONS_LIB}/svg`,
+      deprecatedLib: `${PATH_ICONS_LIB}/raw`, // will be removed later
+    },
+    png: {
+      src: `${ICONS_ROOT}/png`,
+      lib: `${PATH_ICONS_LIB}/png`,
     },
     react: {
-      input: `${ICONS_ROOT}/svg`,
+      src: `${ICONS_ROOT}/svg`,
+      lib: `${PATH_ICONS_LIB}/react`,
       template: `${ICONS_ROOT}/templates/withFdsIconWrapper.jsx`,
-      output: `${PATH_ICONS_LIB}/react`,
     },
     docs: {
-      input: `${PATH_ICONS_LIB}/raw`,
+      input: `${PATH_ICONS_LIB}/svg`,
       template: `${ICONS_ROOT}/templates/docs.hbs`,
       assets: `${ICONS_ROOT}/templates/assets/`,
       css: `${ICONS_ROOT}/style/icons.css`,