+++ CHANGELOG.md
@@ -39,6 +39,24 @@ This is an example of a brief overview of the _Major_ or _Minor_ version changes
 #### `Flex` Component
 - `wrap` prop is no longer accepted on `Flex`
 
+## [8.5] PNG Icons
+
+- Adds `lib/icons/png` PNG distribution for icons
+
+### **Deprecations**
+
+- `lib/icons/raw` is now deprecated. It will be removed in a future major release.
+  please use `lib/icons/svg` when importing raw SVG files.
+
+## [8.4] Add delay to `Tooltip`
+
+- Adds support for `delay` prop in base `Popover` component
+- Adds a `1000`ms hover delay to `Tooltip`
+
+## [8.3] DateInput date formatting
+
+- Adds `dateFormat` prop to `DateInput` to allow different date order (MDY, DMY, YMD)
+
 ## [8.2] DateInput rest props and IconButton role
 
 - Adds rest prop spread onto the Date Input to allow input specific props